Move ExplainabilityMetrics class to metrics package
