+++ explainability-integrationtests/explainability-integrationtests-opennlp/src/test/java/org/kie/trustyai/explainability/integrationtests/opennlp/OpenNLPLimeExplainerTest.java
@@ -51,7 +51,7 @@ import org.kie.trustyai.explainability.model.SimplePrediction;
 import org.kie.trustyai.explainability.model.Type;
 import org.kie.trustyai.explainability.model.Value;
 import org.kie.trustyai.explainability.utils.DataUtils;
-import org.kie.trustyai.explainability.utils.ExplainabilityMetrics;
+import org.kie.trustyai.explainability.metrics.ExplainabilityMetrics;
 import org.kie.trustyai.explainability.utils.ValidationUtils;
 
 import opennlp.tools.langdetect.Language;
